# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Agent Data Gateway",
    "description": "Quant Agent 执行引擎 HTTP/WS/SSE 契约（BE-19）。\n\n## 统一响应\n\n```json\n{\"code\": 0, \"msg\": \"ok\", \"data\": {}, \"ts\": 1719475200000}\n```\n\n- `code=0` 成功；非零见错误码（`docs/10` §1.4）\n- 响应头 `X-Trace-Id` 用于全链路追踪（BE-10）\n\n## 机器可读契约\n\n导出产物：`docs/openapi.json`（`python scripts/export_openapi.py`）。\n人工互校文档：`docs/10. API接口规范.md`。",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/oms/kill_switch": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Trigger Kill Switch",
        "description": "【全局熔断 Kill Switch】\n瞬间阻断所有实盘 Bot 的进程信号并下达市价全平指令给券商网关",
        "operationId": "trigger_kill_switch_api_v1_oms_kill_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchReq"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/orders/{order_id}/cancel": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Cancel Order",
        "description": "带【防并发幂等性锁】的撤单接口 (OMS-03: 同步更新 DB 状态, OMS-12: 审计日志)",
        "operationId": "cancel_order_api_v1_oms_orders__order_id__cancel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderReq"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/orders/{order_id}/modify": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Modify Order",
        "description": "修改订单价格 (改单) 接口 (OMS-03: 同步更新 DB, OMS-12: 审计日志)",
        "operationId": "modify_order_api_v1_oms_orders__order_id__modify_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyOrderReq"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/positions": {
      "get": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "当前持仓",
        "description": "OMS-04: 获取 Redis 缓存中的真实持仓列表",
        "operationId": "get_real_positions_api_v1_oms_positions_get",
        "parameters": [
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "HK",
              "title": "Market"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/bots/{bot_id}/pause": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Pause Bot",
        "description": "OMS-05: 暂停 Bot 算力节点",
        "operationId": "pause_bot_api_v1_oms_bots__bot_id__pause_post",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/bots/{bot_id}/resume": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Resume Bot",
        "description": "OMS-05: 恢复 Bot 算力节点",
        "operationId": "resume_bot_api_v1_oms_bots__bot_id__resume_post",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/bots/{bot_id}/stop": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Stop Bot",
        "description": "OMS-05: 终止 Bot 算力节点",
        "operationId": "stop_bot_api_v1_oms_bots__bot_id__stop_post",
        "parameters": [
          {
            "name": "bot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/algo/start": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Start Algo Order",
        "description": "OMS-08: 接收并启动前端下发的算法拆单任务 (TWAP/VWAP/ICEBERG)\n通过 algo_engine 真实执行拆单逻辑",
        "operationId": "start_algo_order_api_v1_oms_algo_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlgoOrderReq"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/algo/{algo_id}/pause": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Pause Algo Order",
        "description": "OMS-08: 暂停算法拆单",
        "operationId": "pause_algo_order_api_v1_oms_algo__algo_id__pause_post",
        "parameters": [
          {
            "name": "algo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Algo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/algo/{algo_id}/resume": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Resume Algo Order",
        "description": "OMS-08: 恢复算法拆单",
        "operationId": "resume_algo_order_api_v1_oms_algo__algo_id__resume_post",
        "parameters": [
          {
            "name": "algo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Algo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/oms/algo/{algo_id}/cancel": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Cancel Algo Order",
        "description": "OMS-08: 取消算法拆单",
        "operationId": "cancel_algo_order_api_v1_oms_algo__algo_id__cancel_post",
        "parameters": [
          {
            "name": "algo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Algo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/oms/mode/switch": {
      "post": {
        "tags": [
          "OMS & Live Bots"
        ],
        "summary": "Switch Trading Mode",
        "description": "OMS-11: 热切换交易模式 (SANDBOX/LIVE)\n写入 Redis 后立即生效，所有交易逻辑实时读取新模式",
        "operationId": "switch_trading_mode_api_v1_oms_mode_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeSwitchReq"
              },
              "example": {
                "_note": "见 Schema 字段说明；统一响应见 {code,msg,data,ts}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "code": 0,
                  "msg": "ok",
                  "data": {},
                  "ts": 1719475200000
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {
                  "code": 2001,
                  "msg": "请求参数校验失败",
                  "data": null,
                  "ts": 1719475200000,
                  "trace_id": "abcdef0123456789abcdef0123456789"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelOrderReq": {
        "properties": {
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "idempotency_key"
        ],
        "title": "CancelOrderReq"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KillSwitchReq": {
        "properties": {
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "timestamp"
        ],
        "title": "KillSwitchReq"
      },
      "ModeSwitchReq": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ModeSwitchReq"
      },
      "ModifyOrderReq": {
        "properties": {
          "price": {
            "type": "number",
            "title": "Price"
          }
        },
        "type": "object",
        "required": [
          "price"
        ],
        "title": "ModifyOrderReq"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}